refac(utils): reorganized redux utils
also updated imports in some files
