+++ src/components/form/PokeTypeField/PokeTypeField.tsx
@@ -6,10 +6,9 @@ import { BaseButton, Tooltip } from '@showdex/components/ui';
 import { PokemonTypes } from '@showdex/consts/pokemon';
 import { useColorScheme } from '@showdex/redux/store';
 import { formatId } from '@showdex/utils/app';
-import { flattenAlts } from '@showdex/utils/battle';
 import { useUserAgent } from '@showdex/utils/hooks';
 import { percentage } from '@showdex/utils/humanize';
-import { sortUsageAlts } from '@showdex/utils/redux';
+import { flattenAlts, sortUsageAlts } from '@showdex/utils/presets';
 import type { FieldRenderProps } from 'react-final-form';
 import type { ButtonElement, TooltipProps } from '@showdex/components/ui';
 import type { CalcdexPokemonUsageAlt } from '@showdex/redux/store';
