+++ src/redux/services/index.ts
@@ -2,3 +2,4 @@ export * from './actionApi';
 export * from './pkmnApi';
 export * from './presetApi';
 export * from './showdownApi';
+export * from './tagTypes';
