+++ src/redux/services/pkmnApi.ts
@@ -1,6 +1,6 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
-import { PokemonReduxTagType } from '@showdex/consts/redux';
 import { env } from '@showdex/utils/core';
+import { PokemonReduxTagType } from './tagTypes';
 
 /**
  * Serves as the base query for all calls to the pkmn API.
