+++ src/redux/transformers/transformFormatPresetResponse.ts
