+++ src/redux/transformers/transformRandomsPresetResponse.ts
