+++ src/utils/battle/index.ts
@@ -14,9 +14,7 @@ export * from './detectPokemonIdent';
 export * from './detectSpeciesForme';
 export * from './detectStatBoostDelta';
 export * from './detectToggledAbility';
-export * from './detectUsageAlt';
 export * from './findHighestStat';
-export * from './flattenAlts';
 export * from './formatDexDescription';
 export * from './formatStatBoost';
 export * from './getDexForFormat';
@@ -47,7 +45,5 @@ export * from './syncField';
 export * from './syncPokemon';
 export * from './toggleableAbility';
 export * from './toggleRuinAbilities';
-export * from './usageAltPercentFinder';
-export * from './usageAltPercentSorter';
 export * from './usedDynamax';
 export * from './usedTerastallization';
