+++ src/utils/presets/flattenAlts.ts
